pyproject.toml
[tool.poetry]
name = "cats-python"
version = "5.0.0"
description = "Cifrazia Action Transport System for Python"
authors = ["Bogdan Parfenov <adam.brian.bright@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://adambrianbright.github.io/cats-python/"
repository = "https://github.com/AdamBrianBright/cats-python"
documentation = "https://adambrianbright.github.io/cats-python/"
keywords = ["cifrazia", "tcp", "rpc", "cats", "tornado"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Natural Language :: Russian",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: POSIX :: Linux",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]
packages = [
    { include = "cats" }
]

[tool.poetry.urls]
"ChangeLog" = "https://adambrianbright.github.io/cats-python/changelog"
"Bug Tracker" = "https://github.com/AdamBrianBright/cats-python/issues"
"Contact Author" = "https://vk.com/adam_bright"

[tool.poetry.dependencies]
python = ">=3.10.0, <4.0"

richerr = "^0.2.1"
struct-model-python = "^0.1.0"

pytz = "^2021.3"
sentry-sdk = "^1.4.3"
tornado = "^6.1"
ujson = "^4.2.0"
PyYAML = "^5.4.1"
msgpack = "^1.0.2"
Django = { version = ">=2.0, <4.0", optional = true }
djangorestframework = { version = ">=3.12.4", optional = true }
pydantic = { version = ">=1.8.2", optional = true }
djantic = { version = ">=0.3.3", optional = true }
cryptography = "^35.0.0"

[tool.poetry.extras]
django = ["Django", "djangorestframework"]
pydantic = ["pydantic"]
djantic = ["djantic", "pydantic", "Django"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-asyncio = "^0.15.1"
pydantic = "^1.8.2"
pytest-django = "^4.4.0"
pytest-mock = "^3.6.1"
coverage = "^6.0.1"

[tool.poetry.plugins."pytest11"]
"pytest_cats" = "cats.test_utils.pytest.plugin"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
